requirements.txt
geojson==3.1.0
requests==2.28.2
GitPython==3.1.43
pandas==2.2
